refactor: BaseJpaEntity 제거 및 JpaAggregate 단순화 - AggregateRoot 직접 상속으로 구조 개선
JPA 영속성 계층의 상속 구조를 단순화하여 도메인 모델과의 일관성을 향상시켰습니다.

주요 변경 사항:
- BaseJpaEntity 클래스 제거 (불필요한 중간 계층 제거)
- JpaAggregate가 AggregateRoot를 직접 상속하도록 변경
- JPA Auditing 필드(createdAt, updatedAt, version, deletedAt)를 JpaAggregate에 직접 구현
- Hibernate @SoftDelete 어노테이션으로 Soft Delete 처리 (deleted 필드 자동 관리)
- 도메인 이벤트 관리 로직은 AggregateRoot에서 상속받아 사용
- JpaBaseRepository의 제네릭 타입을 BaseJpaEntity에서 JpaAggregate로 변경

테스트 개선:
- TestDomainFixtures.kt로 테스트 Fixture 통합 (TestValueObjects.kt 제거)
- BaseJpaEntityTest 제거 (JpaAggregateTest로 커버됨)
- Soft Delete 테스트 개선 (Hibernate 어노테이션 기반)
- 테스트 구조 명확화 및 중복 제거
